правки кода

+++ app/src/main/java/potapeyko/rss/activities/DrawerItemClickListener.java
@@ -1,6 +1,7 @@
 package potapeyko.rss.activities;
 
-import android.content.Intent;
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
 import android.support.v4.widget.DrawerLayout;
 import android.view.View;
 import android.widget.AdapterView;
@@ -13,6 +14,13 @@ final class DrawerItemClickListener implements ListView.OnItemClickListener {
     private final ListView drawerList;
     private final MainActivity mainActivity;
 
+    private static final int ADD_CHANNEL_ITEM = 0;
+    private static final int CHANGE_CHANNEL_ITEM = 1;
+    private static final int DELETE_CHANNEL_ITEM = 2;
+    private static final int SETTINGS_ITEM = 3;
+    private static final int ABOUT_ITEM = 4;
+
+
     DrawerItemClickListener(DrawerLayout drawerLayout, ListView drawerList, MainActivity mainActivity) {
         this.drawerLayout = drawerLayout;
         this.drawerList = drawerList;
@@ -24,47 +32,46 @@ final class DrawerItemClickListener implements ListView.OnItemClickListener {
         selectItem(position);
     }
 
-    private void selectItem(int position) {
+    private void selectItem(final int position) {
         drawerLayout.closeDrawer(drawerList);
 
         switch (position) {
-            case 0: {
+            case ADD_CHANNEL_ITEM: {
                 NewChanelActivityMy.start(mainActivity);
                 break;
             }
-            case 1: {
-                ChanelChangeActivityMy.start(mainActivity);
+            case CHANGE_CHANNEL_ITEM: {
+                ChannelChangeActivity.start(mainActivity);
                 break;
             }
-            case 2: {
+            case DELETE_CHANNEL_ITEM: {
                 DB db = null;
-                if (mainActivity != null) {
+                if (mainActivity != null && mainActivity.getChanelId()!=-1) {
                     try {
                         db = new DB(mainActivity);
-
                         db.open();
                         db.deleteChanelById(mainActivity.getChanelId());
-
-
-                    } catch (Exception r) {
+                        final SharedPreferences sPref = PreferenceManager.getDefaultSharedPreferences(mainActivity);
+                        final SharedPreferences.Editor ed = sPref.edit();
+                        ed.putLong(MainActivity.CHANEL_ID, -1);
+                        ed.apply();
+                    } catch (Throwable r) {
                         r.printStackTrace();
                     } finally {
                         if (db != null) {
                             db.close();
-                            Intent intent = new Intent(mainActivity, MainActivity.class);
-                            mainActivity.startActivity(intent);
-                            mainActivity.finish();
                         }
+                        ChannelChangeActivity.start(mainActivity);
                     }
                 }
                 break;
             }
 
-            case 3: {
+            case SETTINGS_ITEM: {
                 SettingsActivity.start(mainActivity);
                 break;
             }
-            case 4: {
+            case ABOUT_ITEM: {
                 AboutAppActivityMy.start(mainActivity);
                 break;
             }
